chore: migrate to esm versionrc configs
