--- .versionrc.alpha.cjs
@@ -1,6 +1,0 @@
-const base = require('./.versionrc.base.cjs');
-
-module.exports = {
-  ...base,
-  prerelease: 'alpha'
-};
