+++ .versionrc.alpha.mjs
@@ -1,0 +1,9 @@
+import base from './.versionrc.base.mjs';
+
+
+const config = {
+  ...base,
+  prerelease: 'alpha'
+};
+
+export default config;
